// src/utils/enums/color.enum.ts
export enum COLOR {
  BLACK = "#02020A",
  GREEN = "#31E981",
  GREEN_10 = "rgba(49, 233, 123, .1)",

  GREEN_20 = "rgba(49, 233, 123, .2)",

  GREEN_70 = "rgba(49, 233, 123, .7)",

  WHITE = "#fff",
  WHITE_SECONDARY = "#FFB",
  WHITE_05 = "rgba(255, 255, 255, 0.05)",
  WHITE_10 = "rgba(255, 255, 255, 0.1)",
  WHITE_50 = "rgba(255, 255, 255, 0.5)",
  WHITE_75 = "rgba(255, 255, 255, 0.75)",
  RED = "#FF674D",
  RED_10 = "rgba(255, 103, 77, .1)",

  RED_20 = "rgba(255, 103, 77, .2)",

  RED_70 = "rgba(255, 103, 77, .7)",

  BLUE = "#00A6FA",
  BLUE_HOVER = "rgb(37,179,255)",
  BLUE_ACTIVE = "rgb(75,195,255)",
}